Fix transtypage (autoboxing) Long
Initialize a long with a "L" added to the number

+++ parkingsystem/src/test/java/com/parkit/parkingsystem/FareCalculatorServiceTest.java
@@ -57,7 +57,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareCarWithLessThan30MinParkingTime(int arg){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - (arg * 60 * 1000) );
+        inTime.setTime( System.currentTimeMillis() - (arg * 60 * 1000L) );
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.CAR,false);
         ticket.setInTime(inTime);
@@ -76,7 +76,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareBikeWithLessThan30MinParkingTime(int arg){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - (arg * 60 * 1000) );
+        inTime.setTime( System.currentTimeMillis() - (arg * 60 * 1000L) );
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.BIKE,false);
         ticket.setInTime(inTime);
@@ -95,7 +95,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareCarWithLessThanOneHourParkingTime(int minuteTime, double priceRatio){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - (  minuteTime * 60 * 1000) );//45 minutes parking time should give 3/4th parking fare
+        inTime.setTime( System.currentTimeMillis() - (  minuteTime * 60 * 1000L) );//45 minutes parking time should give 3/4th parking fare
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.CAR,false);
         ticket.setInTime(inTime);
@@ -116,7 +116,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareBikeWithLessThanOneHourParkingTime(int minuteTime, double priceRatio){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - ( minuteTime  * 60 * 1000) );//45 minutes parking time should give 3/4th parking fare
+        inTime.setTime( System.currentTimeMillis() - ( minuteTime  * 60 * 1000L) );//45 minutes parking time should give 3/4th parking fare
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.BIKE,false);
         ticket.setInTime(inTime);
@@ -137,7 +137,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareCar(int minuteTime, double priceRatio){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - (  minuteTime * 60 * 1000) );
+        inTime.setTime( System.currentTimeMillis() - (  minuteTime * 60 * 1000L) );
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.CAR,false);
         ticket.setInTime(inTime);
@@ -157,7 +157,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareBike(int minuteTime, double priceRatio){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - (  minuteTime * 60 * 1000) );
+        inTime.setTime( System.currentTimeMillis() - (  minuteTime * 60 * 1000L) );
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.BIKE,false);
         ticket.setInTime(inTime);
@@ -174,10 +174,10 @@ public class FareCalculatorServiceTest {
     @Order(7)
     @ParameterizedTest(name = "For {0} Day(s) Car Parking Time")
     @CsvSource({"1, 24", "2, 48", "3, 72" })
-    public void calculateFareCarWithMoreThanADayParkingTime(int dayTime, double priceRatio){
+    public void calculateFareCarWithMoreThanADayParkingTime(long dayTime, double priceRatio){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - ((24 * dayTime) * 60 * 60 * 1000));//24 hours parking time should give 24 * parking fare per hour
+        inTime.setTime( System.currentTimeMillis() - ((24 * dayTime) * 60 * 60 * 1000L));//24 hours parking time should give 24 * parking fare per hour
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.CAR,false);
         ticket.setInTime(inTime);
@@ -194,10 +194,10 @@ public class FareCalculatorServiceTest {
     @Order(8)
     @ParameterizedTest(name = "For {0} Day(s) Bike Parking Time")
     @CsvSource({"1, 24", "2, 48", "3, 72" })
-    public void calculateFareBikeWithMoreThanADayParkingTime(int dayTime, double priceRatio){
+    public void calculateFareBikeWithMoreThanADayParkingTime(long dayTime, double priceRatio){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - ((24 * dayTime) * 60 * 60 * 1000));//24 hours parking time should give 24 * parking fare per hour
+        inTime.setTime( System.currentTimeMillis() - ((24 * dayTime) * 60 * 60 * 1000L));//24 hours parking time should give 24 * parking fare per hour
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.BIKE,false);
         ticket.setInTime(inTime);
@@ -215,7 +215,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareUnknownType(){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - (  60 * 60 * 1000) );
+        inTime.setTime( System.currentTimeMillis() - (  60 * 60 * 1000L) );
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, null,false);
         ticket.setInTime(inTime);
@@ -230,7 +230,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareBikeWithFutureInTime(){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() + (  60 * 60 * 1000) );
+        inTime.setTime( System.currentTimeMillis() + (  60 * 60 * 1000L) );
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.BIKE,false);
         ticket.setInTime(inTime);
@@ -247,7 +247,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareCarWithDiscountRecurringVehicule(){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - ( 60 * 60 * 1000) );
+        inTime.setTime( System.currentTimeMillis() - ( 60 * 60 * 1000L) );
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.CAR,false);
 
@@ -271,7 +271,7 @@ public class FareCalculatorServiceTest {
     public void calculateFareBikeWithDiscountRecurringVehicule(){
         //GIVEN
         Date inTime = new Date();
-        inTime.setTime( System.currentTimeMillis() - ( 60 * 60 * 1000) );
+        inTime.setTime( System.currentTimeMillis() - ( 60 * 60 * 1000L) );
         Date outTime = new Date();
         ParkingSpot parkingSpot = new ParkingSpot(1, ParkingType.BIKE,false);
